core/tests/unit_tests/orchestrator/test_hierarchical_orchestrator.py
import unittest
from core.orchestrator.hierarchical_orchestrator import HierarchicalOrchestrator
from unittest.mock import patch, MagicMock

class TestHierarchicalOrchestrator(unittest.TestCase):
    def setUp(self):
        # Set up an instance of the HierarchicalOrchestrator before each test
        self.orchestrator = HierarchicalOrchestrator()

    @patch("core.orchestrator.hierarchical_orchestrator.HierarchicalOrchestrator.coordinate_edge_cloud")
    def test_coordinate_edge_cloud(self, mock_coordinate_edge_cloud):
        # Test coordination between edge and cloud levels
        mock_coordinate_edge_cloud.return_value = "Edge-cloud coordination successful"
        result = self.orchestrator.coordinate_edge_cloud("edge_1", "cloud_1")
        mock_coordinate_edge_cloud.assert_called_once_with("edge_1", "cloud_1")
        self.assertEqual(result, "Edge-cloud coordination successful")

    @patch("core.orchestrator.hierarchical_orchestrator.HierarchicalOrchestrator.failover_to_backup")
    def test_failover_to_backup(self, mock_failover_to_backup):
        # Test failover mechanism to backup orchestrator
        mock_failover_to_backup.return_value = "Failover to backup orchestrator successful"
        result = self.orchestrator.failover_to_backup()
        mock_failover_to_backup.assert_called_once()
        self.assertEqual(result, "Failover to backup orchestrator successful")

    @patch("core.orchestrator.hierarchical_orchestrator.HierarchicalOrchestrator.monitor_edge_clusters")
    def test_monitor_edge_clusters(self, mock_monitor_edge_clusters):
        # Test monitoring of multiple edge clusters
        mock_monitor_edge_clusters.return_value = {"edge_1": "healthy", "edge_2": "degraded"}
        cluster_status = self.orchestrator.monitor_edge_clusters(["edge_1", "edge_2"])
        mock_monitor_edge_clusters.assert_called_once_with(["edge_1", "edge_2"])
        self.assertEqual(cluster_status, {"edge_1": "healthy", "edge_2": "degraded"})

    def tearDown(self):
        # Clean up after each test
        self.orchestrator = None

if __name__ == "__main__":
    unittest.main()
